- Use `child.constructor` to tests for built in types: `String`, `Number` and `Array`
- Reordered to `return` early for all the types above.
- Identify instances of `Node` by checking the built in [`Node#nodeType`](https://developer.mozilla.org/en-US/docs/Web/API/Node/nodeType) integer instead of using `instanceof`, avoiding traversing the prototype inheritance chain.
- Super minor: moved `mounted` / `remounted` calls behind the check for `child` being a bare `Node`. Benefits are close to none, but the cost in complexity is zero, so why not.

Code tested:

```js
div(
    h1({ 'class': 'frzr' }, 'Hello ', b('FRZR'), '!'),
    p('Bacon ipsum dolor amet meatloaf meatball shank porchetta \
        picanha bresaola short loin short ribs capicola fatback beef \
        ribs corned beef ham hock.')
);
```

Before:
```
Benching FRZR <div> with multiple child nodes
29960ns per iteration (33377 ops/sec)
```

After:
```
Benching FRZR <div> with multiple child nodes
13010ns per iteration (76862 ops/sec)
```

VanillaJS with `innerHTML` to compare:
```js
var div = document.createElement('div');
div.innerHTML = '<h1 class="frzr">Hello <b>FRZR</b>!</h1>\
<p>\
    Bacon ipsum dolor amet meatloaf meatball shank porchetta \
    picanha bresaola short loin short ribs capicola fatback beef \
    ribs corned beef ham hock.\
</p>';
```

```
Benching Vanilla JS innerHTML
build.js:19 - 23341ns per iteration (42843 ops/sec)
```

Tried to not add any amount of code that isn't strictly necessary, there is +9 lines but some of those are whitespace :).

![](https://38.media.tumblr.com/f3a414cd6c377f8f3642c772379b93b6/tumblr_myolzvCIxu1sq0a3ho1_500.gif)