@maciejhirsz 2016-08-24
Also #35