@pakastin 2016-08-25
Good stuff! That trick of checking the constructor is especially brilliant. 👍